requirements.txt
RPi.GPIO~=0.7.0
CASlibrary==0.8.0